update flash_attention's invocation with dtype=torch.bfloat16

+++ mindspore-lite/lite_boost/python/model/wan2_1/model.py
@@ -52,7 +52,7 @@ def usp_attn_forward(self,
                      seq_lens,
                      grid_sizes,
                      freqs,
-                     dtype=torch.bfloat16):  # pylint: disable=unused-argument
+                     dtype=torch.bfloat16):
     """Ulysses Sequence Parallel self-attention forward."""
     b, s, n, d = *x.shape[:2], self.num_heads, self.head_dim
     sp_size = dist.get_world_size()
@@ -88,7 +88,8 @@ def usp_attn_forward(self,
         v=v,
         q_lens=seq_lens,
         k_lens=seq_lens,
-        window_size=self.window_size)
+        window_size=self.window_size,
+        dtype=dtype)
 
     # flash_attention now always returns 4D [B, L_q, N//sp_size, D]
